backend/datacollector-service/pyproject.toml
[project]
name = "amber-home"
version = "0.1.0"
description = "Simple home energy data collection and analysis for Amber Electric"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit",
    "amberelectric==2.0.12",
    "pandas",
    "plotly",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
]
